// shared/nextjs/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        FileResource: {
            /** @enum {string} */
            type: "files";
            id: string;
            attributes: components["schemas"]["FileAttributes"];
            relationships: components["schemas"]["FileRelationships"];
            meta?: components["schemas"]["FileMeta"];
        };
// ...
        PostDocument: {
            data: components["schemas"]["PostResource"];
            included?: (components["schemas"]["UserPublicResource"] | components["schemas"]["FileResource"])[];
        };
// ...
        RegistrationCreateAttributes: {
            /** Format: email */
            email: string;
            password: string;
            name: string;
            /** @description 생략하면 Accept-Language로 정한다. */
            locale?: components["schemas"]["Locale"];
        };
// ...
        UserPublicResource: {
            /** @enum {string} */
            type: "users";
            id: string;
            attributes: components["schemas"]["UserPublicAttributes"];
            relationships: components["schemas"]["UserPublicRelationships"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}